requirements.txt
streamlit>=1.28.0
streamlit-webrtc>=0.47.0
opencv-python>=4.8.0
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
Pillow>=9.5.0
numpy>=1.24.0
av>=10.0.0